Revert "Project and seeders"
This reverts commit 21aaa418f70870409e130322e267abd60dc3c9a6.

+++ database/migrations/2021_03_07_111811_create_stories_table.php
@@ -15,7 +15,7 @@ class CreateStoriesTable extends Migration
     {
         Schema::create('stories', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('project_id')->nullable();
+            $table->integer('project_id')->nullable();
             $table->string('title');
             $table->text('description');
             $table->text('tests');
